Merge!

+++ src/Algorithms/Operation.java
@@ -1,7 +1,9 @@
 package Algorithms;
 
+import java.math.BigInteger;
+
 public interface Operation {
-	public int execute(int element1 , int element2);
+	public BigInteger execute(BigInteger element1 , BigInteger element2);
 	public boolean isOperator(char operator);
 	public Operation clone();
 }
